+++ verticapy/io/parsers/csv.py
@@ -29,6 +29,7 @@ from verticapy.utils._toolbox import *
 from verticapy.errors import ExtensionError, ParameterError, MissingRelation
 from ..flex import compute_flextable_keys
 from verticapy.io.sql.utils._format import format_schema_table, clean_query
+from verticapy.io.parsers._utils import extract_compression, get_first_file
 
 
 def guess_sep(file_str: str):
