+++ verticapy/io/parsers/json.py
@@ -24,6 +24,7 @@ from verticapy.utils._toolbox import *
 from verticapy.errors import ExtensionError, ParameterError, MissingRelation
 from ..flex import compute_flextable_keys
 from verticapy.io.sql.utils._format import quote_ident, format_schema_table, clean_query
+from verticapy.io.parsers._utils import extract_compression
 
 
 def pjson(path: str, ingest_local: bool = True):
